new version of task assign
Integrated the new version with an old one. Now we will be able to set specific actions to happen when you complete a task for the first time. Probably use this to see wheather the player has completed all of the tasks and if not give penalties in the calculation of the final score at the end of the mission